+++ src/backend/joanie/tests/core/test_api_contract.py
@@ -129,6 +129,68 @@ class ContractApiTest(BaseAPITestCase):
             },
         )
 
+    def test_api_contracts_list_filter_is_signed(self):
+        """
+        Authenticated user can query owned contracts and filter them by signature state.
+        """
+        user = factories.UserFactory()
+        token = self.generate_token_from_user(user)
+
+        unsigned_contracts = factories.ContractFactory.create_batch(
+            5, order__owner=user
+        )
+
+        signed_contract = factories.ContractFactory.create(
+            order__owner=user,
+            signed_on=timezone.now(),
+            definition_checksum="test",
+            context={"title": "test"},
+        )
+
+        # Create random contracts that should not be returned
+        factories.ContractFactory.create_batch(5)
+
+        # - List without filter should return 6 contracts
+        with self.assertNumQueries(2):
+            response = self.client.get(
+                "/api/v1.0/contracts/",
+                HTTP_AUTHORIZATION=f"Bearer {token}",
+            )
+
+        self.assertEqual(response.status_code, 200)
+        content = response.json()
+        self.assertEqual(content["count"], 6)
+
+        # - Filter by is_signed=false should return 5 contracts
+        with self.assertNumQueries(2):
+            response = self.client.get(
+                "/api/v1.0/contracts/?is_signed=false",
+                HTTP_AUTHORIZATION=f"Bearer {token}",
+            )
+
+        self.assertEqual(response.status_code, 200)
+        content = response.json()
+        count = content["count"]
+        result_ids = [result["id"] for result in content["results"]]
+        self.assertEqual(count, 5)
+        self.assertCountEqual(
+            result_ids, [str(contract.id) for contract in unsigned_contracts]
+        )
+
+        # - Filter by is_signed=true should return 1 contract
+        with self.assertNumQueries(2):
+            response = self.client.get(
+                "/api/v1.0/contracts/?is_signed=true",
+                HTTP_AUTHORIZATION=f"Bearer {token}",
+            )
+
+        self.assertEqual(response.status_code, 200)
+        content = response.json()
+        count = content["count"]
+        result_ids = [result["id"] for result in content["results"]]
+        self.assertEqual(count, 1)
+        self.assertEqual(result_ids, [str(signed_contract.id)])
+
     def test_api_contracts_retrieve_anonymous(self):
         """
         Anonymous user cannot query a contract.